+++ src/collector/CollectorWrapper.test.ts
@@ -14,6 +14,8 @@ import SessionIdHandler from '../session-id-handler/SessionIdHandler'
 import { nop } from '../utils/nop'
 import TestNameHandler from '../test-name-handler/TestNameHandler'
 import SessionStorageTestNameHandler from '../test-name-handler/SessionStorageTestNameHandler'
+import completeOptions from '../utils/completeOptions'
+import SessionTraitHandler from '../session-trait/SessionTraitHandler'
 
 describe('CollectorWrapper', () => {
   beforeEach(() => {
@@ -115,6 +117,20 @@ describe('CollectorWrapper', () => {
 
         expect(KeyDownEventListener.prototype.init).toHaveBeenCalledOnce()
       })
+    })
+  })
+
+  describe('identifySession', () => {
+    it('delegates session trait to handler', () => {
+      const collectorWrapper = new CollectorWrapper(
+        completeOptions({ debug: true }),
+        global.window,
+        new MemorySessionIdHandler(),
+        new SessionStorageTestNameHandler(),
+      )
+      const mock = vi.spyOn(SessionTraitHandler.prototype, 'handle').mockImplementation(nop)
+      collectorWrapper.identifySession('connected', true)
+      expect(mock).toHaveBeenCalledWith('connected', true)
     })
   })
 })