+++ src/collector/GravityCollector.ts
@@ -1,5 +1,5 @@
 import CollectorWrapper from './CollectorWrapper'
-import { CollectorOptions } from '../types'
+import { CollectorOptions, TraitValue } from '../types'
 import windowExists from '../utils/windowExists'
 import completeOptions from '../utils/completeOptions'
 import SessionStorageSessionIdHandler from '../session-id-handler/SessionStorageSessionIdHandler'
@@ -28,5 +28,13 @@ export default class GravityCollector {
         new SessionStorageTestNameHandler(),
       ),
     )
+  }
+
+  static identifySession(traitName: string, traitValue: TraitValue) {
+    const collectorWrapper = (window as any)._GravityCollector.collectorWrapper
+    if (collectorWrapper === undefined) {
+      throw new Error('Gravity Data Collector was not initialized : please call window.GravityCollector.init() before')
+    }
+    collectorWrapper.identifySession(traitName, traitValue)
   }
 }