+++ src/user-action/userActionSessionSender.ts
@@ -1,12 +1,7 @@
 import { SessionUserAction } from '../types'
 import crossfetch from 'cross-fetch'
 import { nop } from '../utils/nop'
-
-export const GRAVITY_SERVER_ADDRESS = 'https://smartestinggravityserver.herokuapp.com'
-
-export function buildGravityTrackingApiUrl(authKey: string, gravityServerUrl: string) {
-  return `${gravityServerUrl}/api/tracking/${authKey}/publish`
-}
+import { buildGravityTrackingPublishApiUrl } from '../gravityEndPoints'
 
 export function defaultUserActionSessionSender(
   authKey: string,
@@ -22,10 +17,11 @@ export function debugUserActionSessionSender(maxDelay: number, output: (args: an
   return (sessionActions: SessionUserAction[]) => {
     if (maxDelay === 0) {
       printSessionUserActions(sessionActions, output)
+    } else {
+      setTimeout(() => {
+        printSessionUserActions(sessionActions, output)
+      }, Math.random() * maxDelay)
     }
-    setTimeout(() => {
-      printSessionUserActions(sessionActions, output)
-    }, Math.random() * maxDelay)
   }
 }
 
@@ -51,7 +47,7 @@ export async function sendSessionUserActions(
     if (source !== null) {
       headers.Origin = source
     }
-    const response = await fetch(buildGravityTrackingApiUrl(authKey, gravityServerUrl), {
+    const response = await fetch(buildGravityTrackingPublishApiUrl(authKey, gravityServerUrl), {
       method: 'POST',
       body: JSON.stringify(sessionActions),
       headers,